split up products
